separated main.js to modules, added type:module to package.json
