+++ package.json
@@ -25,5 +25,6 @@
   },
   "engines": {
     "node": ">=12.0.0"
-  }
+  },
+  "type": "module"
 }
